+++ changelog.md
@@ -24,6 +24,8 @@
 #### 2024.06.02 (**0.3.1**):
 * Fixes.
 
-#### 2024.06.29 (**0.3.2**):
+#### 2024.06.29 (**0.4.0**):
 * Minor changes.
-* The Code property has been renamed to Index.
+* The "Code" property has been renamed to "Index".
+* Added: Sheet metal settings: radius, k-factor calculation.
+* Added: Sheet metal part settings: color and density.